### 2022-03-08 #4235 @editorialbot
[PRE REVIEW]: imucal - A Python library to calibrate 6 DOF IMUs
**Submitting author:** @AKuederle (<a href="http://orcid.org/0000-0002-5686-281X">Arne Küderle</a>)
**Repository:** https://github.com/mad-lab-fau/imucal
**Branch with paper.md** (empty if default branch):
**Version:** v2.1.0
