### 2022-03-08 #4235 @AKuederle
Suggestions for reviewers: tuliofalmeida, aaa34169, jizecn, sptennak

The suggestions are based on knowledge in Python and movement analysis or wireless sensor tech.
