### 2022-03-08 #4235 @Kevin-Mattheus-Moerman
@AKuederle thanks for this submission. I am the AEiC on call this week for JOSS. As you can see I have started looking for a handling editor.
In the meantime you could perhaps check those potentially invalid DOIs. :point_up: You can update the paper by calling `@editorialbot generate pdf`, and you can check the references by calling `@editorialbot check references`.
