+++ paper/imucal.bib
@@ -1,7 +1,7 @@
 
 @article{Ferraris1994,
   title = {Calibration of Three-Axial Rate Gyros without Angular Velocity Standards},
-  author = {Ferraris, Franco and Gorini, Italo and Grimaldi, Ugo and Parvis, Marco},
+  author = {Ferraris, F. and Gorini, I. and Grimaldi, U. and Parvis, M.},
   year = {1994},
   month = apr,
   journal = {Sensors and Actuators A: Physical},
@@ -21,8 +21,9 @@
   year = {1995},
   journal = {Sensors and Materials},
   volume = {7},
+  number = {5},
   pages = {311--330},
-  doi = {10.1088/0967}
+  url = {https://myukk.org/SM2017/article.php?ss=10210}
 }
 
 @inproceedings{Kozlov2014,
