### 2022-03-08 #4235 @AKuederle
10.1016/0924-42479480031-6:  This seems to be incorrectly extracted from the bot. In the bibtex file and the paper the following doi is listed 10.1016/0924-4247(94)80031-6, which resolves correctly (Notice the brackets around the 94)

10.1088/0967: This DOI is indeed incorrect. However, it seems that no proper DOI exists for this paper. I could find the website corresponding to the original publisher: https://myukk.org/SM2017/article.php?ss=10210  However, it seems like they did not provide a DOI for their papers back in 1995. I removed the DOI from the paper for now and added the link to the respective journal page instead. Is that sufficient? How do you suggest I proceed here?
