### 2022-03-08 #4235 @Kevin-Mattheus-Moerman
@openjournals/dev I've been seeing a lot of these "false" invalid DOIs :point_up: most seem to be due to the brackets. Is this something we can fix in our system?
