### 2022-03-08 #4235 @Kevin-Mattheus-Moerman
@AKuederle thanks for checking that. On the second paper, if it does not have a DOI, so be it, you can leave it out, that is no problem.
