### 2022-03-08 #4235 @xuanxu
> I've been seeing a lot of these "false" invalid DOIs ☝️ most seem to be due to the brackets. Is this something we can fix in our system?

Yes, code validating DOI updated! It should be fixed now
