### 2022-03-28 #4235 @adi3
@arianesasso @briandesilva @tuelwer @LukasAdamowicz @nhejazi @finsberg - Would any of you be interested in reviewing this paper on robotics? We would greatly appreciate your help!
