+++ pyproject.toml
@@ -25,7 +25,7 @@ pandas = "^1.1.3"
 numpy = "^1.19.2"
 matplotlib = {version = "^3.3.2", optional = true}
 h5py = {version = "^2.10.0", optional = true}
-typing-extensions = "^3.7.4"
+typing-extensions = ">=3.7.4"
 packaging = "^21.3"
 
 [tool.poetry.extras]
