+++ pyproject.toml
@@ -38,7 +38,7 @@ pytest = "^6.1.1"
 pytest-cov = "^2.10.1"
 coverage = "^5.3"
 prospector = "^1.7.6"
-black = "^20.8b1"
+black = "^22.3.0"
 sphinx-gallery = "^0.10.1"
 Sphinx = "^4.4.0"
 numpydoc = "^1.2.0"
