+++ pyproject.toml
@@ -20,7 +20,7 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.7"
+python = ">=3.7.1,<3.11"
 pandas = "^1.1.3"
 numpy = "^1.19.2"
 matplotlib = {version = "^3.3.2", optional = true}
