+++ dodo.py
@@ -38,7 +38,13 @@ def update_version_strings(file_path, new_version):
     with open(file_path, "r+") as f:
         content = f.read()
         f.seek(0)
-        f.write(re.sub(version_regex, lambda match: '{}{}"'.format(match.group(1), new_version), content,))
+        f.write(
+            re.sub(
+                version_regex,
+                lambda match: '{}{}"'.format(match.group(1), new_version),
+                content,
+            )
+        )
         f.truncate()
 
 
