+++ .github/workflows/test-and-lint.yml
@@ -23,8 +23,11 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install poetry
+        # We are using the prerelase version to get the "check lock command"
+        pip install poetry="1.2.0b1"
         poetry config virtualenvs.in-project true
+        # Fail the CI if the lock file is not up to date
+        poetry lock --check
         poetry install -E calplot -E h5py
     - name: Testing
       run: |
