### 2022-04-25 #4338 @editorialbot
[REVIEW]: imucal - A Python library to calibrate 6 DOF IMUs
**Submitting author:** @AKuederle (<a href="http://orcid.org/0000-0002-5686-281X">Arne Küderle</a>)
**Repository:** https://github.com/mad-lab-fau/imucal
**Branch with paper.md** (empty if default branch):
**Version:** v2.2.1
**Archive:** 10.17605/OSF.IO/37TD9
