### 2022-04-25 mad-lab-fau/imucal#18 @tuliofalmeida
Review JOSS
This is an issue created as part of [imucal - JOSS Review](https://github.com/openjournals/joss-reviews/issues/4338)

Hi,

I tested the library (pypi version) by running the examples and on own data and it worked fine. The documentation is also well done. I have only one suggestion to expand the scope and use by other laboratories.

Would it be possible to create a version to work in server/Jupyter ? I believe that if this information can be passed through a dictionary based on the index of each moment, in this way it could be integrated with other applications and increase the impact of the work. The user would plot the data and extract the times himself.

Best,

Túlio
