### 2022-04-25 mad-lab-fau/imucal#18 @AKuederle
Hi Túlio,

thanks for the feedback! I am not 100% sure if I understand your suggestion correctly.

1. Would it be possible to create a version in server/Jupyter?
The library works perfectly fine on the server or on Jupyter. Only the GUI works only locally, as it uses a local GUI toolkit. If you have the timings already or recorded the different calibration phases as different recordings you can use [ferraris_regions_from_df](https://imucal.readthedocs.io/en/latest/modules/generated/imucal.ferraris_regions_from_df.html) or [ferraris_regions_from_section_list](https://imucal.readthedocs.io/en/latest/modules/generated/imucal.ferraris_regions_from_section_list.html) (or create some custom way to create a `FerrarisSignalRegions` object) to initialize the calibration.

Then you can use the library just fine in Jupyter or on a server, or in a another library (that is what we are doing internally as well).
Is that what you meant? If yes, I will make sure that it is clear in the README, that you do not need to use the GUI to find the different sections.
