### 2022-04-25 mad-lab-fau/imucal#18 @tuliofalmeida
Perfect, that was exactly it. I had only tested it on my own pc and reading the repository I had understood that it was in general.
I had already given the report of my review. Thanks.
