### 2022-04-25 mad-lab-fau/imucal#18 @AKuederle
Will keep it open until I modified the README accordingly to make that clearer :)
