### 2022-04-26 127054b
Made sure the docu is easily discoverable, and that it is clear that you can use other methods besides the GUI to creat the different sections
