### 2022-04-26 fa59816
Removed 3.10 from the CI again, but added a note about it in the README.
