### 2022-04-26 mad-lab-fau/imucal#18 @AKuederle
I have changed the README in 127054ba22ab211f72e97b868b89596c542b1e2d .

I will close the issue, but feel free to reopen, if you think the change is not sufficient.
