### 2022-04-26 397fef9
Allowed newer version of h5py and updated lock file
