+++ pyproject.toml
@@ -24,7 +24,7 @@ python = ">=3.7.1,<3.11"
 pandas = "^1.1.3"
 numpy = "^1.19.2"
 matplotlib = {version = "^3.3.2", optional = true}
-h5py = {version = "^2.10.0", optional = true}
+h5py = {version = "^3", optional = true}
 typing-extensions = ">=3.7.4"
 packaging = "^21.3"
 
