### 2022-04-26 eb39d78
Fixed code to support new h5py version
