### 2022-04-26 df1d145
Made check for string dtypes more explicit
