+++ imucal/calibration_info.py
@@ -6,6 +6,7 @@ from typing import Tuple, Union, TypeVar, ClassVar, Optional, Type, Iterable
 
 import numpy as np
 import pandas as pd
+from h5py import check_string_dtype
 from packaging.version import Version
 
 CalInfo = TypeVar("CalInfo", bound="CalibrationInfo")
@@ -292,12 +293,13 @@ class CalibrationInfo:
             subcls = cls.find_subclass_from_cal_type(hdf["cal_type"][()].decode("utf-8"))
             data = {}
             for k in fields(subcls):
-                try:
+                dp = hdf.get(k.name)
+                if check_string_dtype(dp.dtype):
                     # String data
-                    data[k.name] = hdf.get(k.name).asstr()[()]
-                except TypeError:
+                    data[k.name] = dp.asstr()[()]
+                else:
                     # No string data
-                    data[k.name] = hdf.get(k.name)[()]
+                    data[k.name] = dp[()]
         return subcls._from_list_dict(data)
 
 
