### 2022-04-26 1f5a9f6
Changed the support statement after the h5py update
