+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) (+ the Migration Guide section), and 
 this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [2.2.2] - 26.04.2022
+# [2.2.0] - 26.04.2022
 
 - Dropped support for `h5py` < 3.0.0. This should only affect users using the hdf5 export feature.
   Dropping old versions allows for proper support of Python 3.10.
