### 2022-04-26 947d12a
Fixed accidantal top-level import of h5py
