+++ docs/requirements.txt
@@ -1,8 +1,0 @@
-sphinx>=4.4.0
-numpydoc>=1.2.0
-sphinx_gallery>=0.10.1
-pydata-sphinx-theme>=0.8.0
-recommonmark>=0.6.0
-memory_profiler>=0.57.0
-matplotlib>=3.3.2
-toml>=0.10.2
