+++ .readthedocs.yml
@@ -10,9 +10,13 @@ sphinx:
   configuration: docs/conf.py
 
 # Optionally set the version of Python and requirements required to build your docs
-python:
-  version: 3.7
-  install:
-    - requirements: docs/requirements.txt
-    - method: pip
-      path: .
+build:
+  os: "ubuntu-20.04"
+    tools:
+      python: "3.8"
+    jobs:
+      pre_install:
+        - pip install poetry
+        - poetry config virtualenvs.in-project true
+      post_build:
+        - poetry install
