+++ .readthedocs.yml
@@ -12,11 +12,11 @@ sphinx:
 # Optionally set the version of Python and requirements required to build your docs
 build:
   os: "ubuntu-20.04"
-    tools:
-      python: "3.8"
-    jobs:
-      pre_install:
-        - pip install poetry
-        - poetry config virtualenvs.in-project true
-      post_build:
-        - poetry install
+  tools:
+    python: "3.8"
+  jobs:
+    pre_install:
+      - pip install poetry
+      - poetry config virtualenvs.in-project true
+    post_build:
+      - poetry install
