+++ .readthedocs.yml
@@ -9,14 +9,13 @@ version: 2
 sphinx:
   configuration: docs/conf.py
 
-# Optionally set the version of Python and requirements required to build your docs
 build:
   os: "ubuntu-20.04"
   tools:
     python: "3.8"
   jobs:
-    pre_install:
+    pre_create_environment:
       - pip install poetry
-      - poetry config virtualenvs.in-project true
-    post_build:
+      - poetry config virtualenvs.create false
+    post_install:
       - poetry install
