### 2022-05-02 cb07352
Updated citation and removed old DOI
