### 2022-05-02 #4338 @AKuederle
Thank you @adi3!

We archived the repo on OSF under the following DOI: DOI 10.17605/OSF.IO/37TD9
The current release version is [v2.2.1](https://github.com/mad-lab-fau/imucal/tree/v2.2.1)

Let us know if you need anything else from us!
