+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "imucal"
-version = "2.2.0"
+version = "2.2.1"
 description = "A Python library to calibrate 6 DOF IMUs"
 authors = [
     "Arne Küderle <arne.kuederle@fau.de>",
