### 2022-05-09 #4338 @AKuederle
@xuanxu Fixed the formatting error in the paper that should have led to the error above (Thanks again!). Should I create a new tag or do I need to do anything else to restart the publication process?
