### 2022-05-09 #4338 @xuanxu
@AKuederle There's no need for a new tag. Now @adi3 can just continue with the editorial process (if the `recommend-accept` command does not return any other error 😅 🤞).
